refactor: split Subclip component
