+++ frontend/src/pages/AssetEditor/Preview.jsx
@@ -13,100 +13,8 @@ import {
   Button,
 } from '/src/components'
 import VideoPlayer from '/src/containers/VideoPlayer'
-
-const SubclipRow = styled.div`
-  display: flex;
-  gap: 8px;
-  padding: 4px;
-  align-items: center;
-  justify-content: center;
-  border-bottom: 1px solid var(--color-border);
-`
-
-const SubclipContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  background-color: var(--color-surface-03);
-  border: 1px solid var(--color-surface-04);
-  width: 100%;
-
-  h3 {
-    font-family: monospace;
-    font-size: 12px;
-    margin: 0;
-    padding: 4px;
-  }
-`
-
-const Subclip = ({
-  index,
-  title,
-  mark_in,
-  mark_out,
-  setSubclips,
-  selection,
-  setSelection,
-}) => {
-  const onSetMarks = (marks) => {
-    setSubclips((subclips) => {
-      const newSubclips = [...subclips]
-      newSubclips[index] = { ...newSubclips[index], ...marks }
-      return newSubclips
-    })
-  }
-
-  const onTitleChange = (e) => {
-    setSubclips((subclips) => {
-      const newSubclips = [...subclips]
-      newSubclips[index] = { ...newSubclips[index], title: e }
-      return newSubclips
-    })
-  }
-
-  const onRemove = () => {
-    setSubclips((subclips) => {
-      const newSubclips = [...subclips]
-      newSubclips.splice(index, 1)
-      return newSubclips
-    })
-  }
-
-  const fps = 25
-  const startTC = new Timecode(mark_in * fps, fps)
-  const endTC = new Timecode(mark_out * fps, fps)
-
-  return (
-    <SubclipContainer>
-      <h3>
-        {startTC.toString()} - {endTC.toString()}
-      </h3>
-      <SubclipRow>
-        <InputText value={title} onChange={onTitleChange} style={{ flex: 1 }} />
-        <Button
-          icon="delete"
-          tooltip="Delete subclip"
-          onClick={() => onRemove()}
-        />
-        <Button
-          icon="download"
-          tooltip="Update subclip from selection"
-          onClick={() => onSetMarks(selection)}
-        />
-        <Button
-          icon="upload"
-          tooltip="Select region"
-          onClick={() =>
-            setSelection({
-              mark_in: mark_in || null,
-              mark_out: mark_out || null,
-            })
-          }
-        />
-      </SubclipRow>
-    </SubclipContainer>
-  )
-}
+import Subclip from './Subclip'
+import { useKeyDown } from '/src/hooks'
 
 const Preview = ({ assetData, setAssetData }) => {
   const accessToken = nebula.getAccessToken()
@@ -169,11 +77,37 @@ const Preview = ({ assetData, setAssetData }) => {
 
   // Actions
 
+  const onNewSubclip = () => {
+    if (!(selection.mark_in && selection.mark_out)) {
+      toast.error('Please select a region first')
+      return
+    }
+
+    if (selection.mark_in >= selection.mark_out) {
+      toast.error('Please select a valid region')
+      return
+    }
+
+    setSubclips((subclips) => [
+      ...subclips,
+      { title: `SubClip ${subclips.length + 1}`, ...selection },
+    ])
+  }
+
+  useKeyDown('v', onNewSubclip)
+
+  /*
   const onNewSubclip = useCallback(() => {
     if (!(selection.mark_in && selection.mark_out)) {
       toast.error('Please select a region first')
       return
     }
+
+    if (selection.mark_in >= selection.mark_out) {
+      toast.error('Please select a valid region')
+      return
+    }
+
     setSubclips((subclips) => [
       ...subclips,
       { title: `SubClip ${subclips.length + 1}`, ...selection },
@@ -189,6 +123,7 @@ const Preview = ({ assetData, setAssetData }) => {
     window.addEventListener('keydown', handleKeyDown)
     return () => window.removeEventListener('keydown', handleKeyDown)
   }, [onNewSubclip])
+   */
 
   // Render
